requirements.txt
comtypes
google-genai
pydantic
python-dotenv
httpx